refactor 01
